Apply suggestions from code review

+++ src/redis_bitmap_string.cc
@@ -45,7 +45,7 @@ rocksdb::Status BitmapString::SetBit(const Slice &ns_key,
   return storage_->Write(rocksdb::WriteOptions(), &batch);
 }
 
-rocksdb::Status BitmapString::BitCount(const std::string &raw_value, int start, int stop, uint32_t *cnt) {
+rocksdb::Status BitmapString::BitCount(const std::string &raw_value, int64_t start, int64_t stop, uint32_t *cnt) {
   *cnt = 0;
   auto string_value = raw_value.substr(STRING_HDR_SIZE, raw_value.size() - STRING_HDR_SIZE);
   /* Convert negative indexes */
@@ -57,12 +57,12 @@ rocksdb::Status BitmapString::BitCount(const std::string &raw_value, int start, 
   if (stop < 0) stop = strlen + stop;
   if (start < 0) start = 0;
   if (stop < 0) stop = 0;
-  if (stop >= static_cast<int>(strlen)) stop = strlen - 1;
+  if (stop >= static_cast<int64_t>(strlen)) stop = strlen - 1;
 
   /* Precondition: end >= 0 && end < strlen, so the only condition where
      * zero can be returned is: start > stop. */
   if (start <= stop) {
-    int bytes = stop - start + 1;
+    int64_t bytes = stop - start + 1;
     *cnt = redisPopcount((unsigned char *) (&string_value[0] + start), bytes);
   }
   return rocksdb::Status::OK();
@@ -70,8 +70,8 @@ rocksdb::Status BitmapString::BitCount(const std::string &raw_value, int start, 
 
 rocksdb::Status BitmapString::BitPos(const std::string &raw_value,
                                      bool bit,
-                                     int start,
-                                     int stop,
+                                     int64_t start,
+                                     int64_t stop,
                                      bool stop_given,
                                      int64_t *pos) {
   auto string_value = raw_value.substr(STRING_HDR_SIZE, raw_value.size() - STRING_HDR_SIZE);
@@ -81,12 +81,12 @@ rocksdb::Status BitmapString::BitPos(const std::string &raw_value,
   if (stop < 0) stop = strlen + stop;
   if (start < 0) start = 0;
   if (stop < 0) stop = 0;
-  if (stop >= static_cast<int>(strlen)) stop = strlen - 1;
+  if (stop >= static_cast<int64_t>(strlen)) stop = strlen - 1;
 
   if (start > stop) {
     *pos = -1;
   } else {
-    int bytes = stop - start + 1;
+    int64_t bytes = stop - start + 1;
     *pos = redisBitpos((unsigned char *) (&string_value[0] + start), bytes, bit);
 
     /* If we are looking for clear bits, and the user specified an exact
@@ -113,7 +113,7 @@ rocksdb::Status BitmapString::BitPos(const std::string &raw_value,
  * This function started out as:
  * https://github.com/antirez/redis/blob/94f2e7f/src/bitops.c#L40
  * */
-size_t BitmapString::redisPopcount(unsigned char *p, int count) {
+size_t BitmapString::redisPopcount(unsigned char *p, int64_t count) {
   size_t bits = 0;
   uint32_t *p4;
   static const unsigned char bitsinbyte[256] =
@@ -185,7 +185,7 @@ size_t BitmapString::redisPopcount(unsigned char *p, int count) {
  * This function started out as:
  * https://github.com/antirez/redis/blob/94f2e7f/src/bitops.c#L101
  * */
-int64_t BitmapString::redisBitpos(unsigned char *c, int count, int bit) {
+int64_t BitmapString::redisBitpos(unsigned char *c, int64_t count, int bit) {
   uint64_t *l;
   uint64_t skipval, word = 0, one;
   int64_t pos = 0; /* Position of bit, to return to the caller. */
