+++ src/redis_bitmap_string.h
@@ -13,11 +13,12 @@ class BitmapString : public Database {
   BitmapString(Engine::Storage *storage, const std::string &ns) : Database(storage, ns) {}
   rocksdb::Status GetBit(const std::string &raw_value, uint32_t offset, bool *bit);
   rocksdb::Status SetBit(const Slice &ns_key, std::string *raw_value, uint32_t offset, bool new_bit, bool *old_bit);
-  rocksdb::Status BitCount(const std::string &raw_value, int start, int stop, uint32_t *cnt);
-  rocksdb::Status BitPos(const std::string &raw_value, bool bit, int start, int stop, bool stop_given, int64_t *pos);
+  rocksdb::Status BitCount(const std::string &raw_value, int64_t start, int64_t stop, uint32_t *cnt);
+  rocksdb::Status BitPos(const std::string &raw_value, bool bit,
+                          int64_t start, int64_t stop, bool stop_given, int64_t *pos);
  private:
-  size_t redisPopcount(unsigned char *p, int count);
-  int64_t redisBitpos(unsigned char *c, int count, int bit);
+  size_t redisPopcount(unsigned char *p, int64_t count);
+  int64_t redisBitpos(unsigned char *c, int64_t count, int bit);
 };
 
 }  // namespace Redis
